requirements.txt
numpy
scipy
pandas
parse
scikit-learn
scikit-image
Pillow
pyTelegramBotAPI
pyttsx3
